feat(termcolor): print Colored Output on pipe
This catches sets fatih/color's NoColor variable for Stderr. By default,
it checks for color support for stdout. It creates issue when we are
piping the command to another command. Because, the pipe itself doesn't
support colors. Thus NoColor is variable is set fro Stderr where the
logger is going to write the output.