fix(cache): set new path to Image for output
This patch includes a fix for setting the proper path when the original
source file is renamed. Previously, if the source path was renamed, the
cache was loaded correctly but the Image/Video still pointed to the old
image path, causing issues when using the Image/Video in template
output.